Check if paths exist before deleting them. Reported by Renaud Richardet.
git-svn-id: https://svn.eu.apache.org/repos/asf/lucene/nutch/trunk@487143 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/nutch/crawl/LinkDb.java
@@ -315,11 +315,13 @@ public class LinkDb extends ToolBase implements Mapper, Reducer {
     FileSystem fs = new JobClient(job).getFs();
     Path old = new Path(linkDb, "old");
     Path current = new Path(linkDb, CURRENT_NAME);
-    fs.delete(old);
-    fs.rename(current, old);
+    if (fs.exists(current)) {
+      if (fs.exists(old)) fs.delete(old);
+      fs.rename(current, old);
+    }
     fs.mkdirs(linkDb);
     fs.rename(newLinkDb, current);
-    fs.delete(old);
+    if (fs.exists(old)) fs.delete(old);
   }
 
   public static void main(String[] args) throws Exception {
